removed createInterceptDownload file and use existing code for http download
